feat(button-group) move tooltip to item-element component
